modify directories for pages router
