+++ src/components/libs/QualificationsList.ts
