+++ src/components/pages/MyProfileScreen/MyProfileScreen.tsx
@@ -1,9 +1,9 @@
-import { Footer } from "./components/Footer";
-import { Header } from "./components/Header";
-import MyProfile from "./components/MyProfile";
-import styles from "../styles/main.module.css";
+import { Header } from "@/components/parts/Header";
+import styles from "../../../styles/main.module.css";
+import MyProfile from "./MyProfile";
+import HomeButton from "@/components/parts/HomeButton";
+import { Footer } from "@/components/parts/Footer";
 // import buttonStyles from "../styles/buttonStyles.module.css"
-import HomeButton from "./components/HomeButton";
 
 const MyProfileScreen = () => {
   return (
