+++ src/components/pages/ProductListScreen/Product.tsx
@@ -1,0 +1,13 @@
+import HomeButton from "@/components/parts/HomeButton";
+import ProductList from "./ProductList";
+
+const Product = () => {
+  return (
+    <>
+      <ProductList />
+      <HomeButton />
+    </>
+  );
+};
+
+export default Product;
