+++ src/components/pages/QualificationsScreen/QualificationsScreen.tsx
@@ -1,10 +1,10 @@
 // import { TitleType } from "@/types/Title";
-import styles from "../styles/main.module.css";
+import HomeButton from "@/components/parts/HomeButton";
+import styles from "../../../styles/main.module.css";
 // import buttonStyles from "../styles/buttonStyles.module.css"
-import { Header } from "./components/Header";
-import { Footer } from "./components/Footer";
-import Qualifications from "./components/Qualifications";
-import HomeButton from "./components/HomeButton";
+import Qualifications from "./Qualifications";
+import { Header } from "@/components/parts/Header";
+import { Footer } from "@/components/parts/Footer";
 
 const QualificationsScreen = (/**props: TitleType*/) => {
   return (
