+++ src/components/parts/Footer.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import frame from "../../../styles/frame.module.css";
+
+export const Footer = () => {
+  return (
+    <footer className={frame.footer}>
+      <p>&copy; 2025 profile site</p>
+    </footer>
+  );
+};
