+++ src/components/parts/Header.tsx
@@ -1,9 +1,10 @@
 import { AppBar, Container, Toolbar, Typography } from "@mui/material";
 import { TitleType } from "@/types/Title";
 // import { useHamburgerIcon } from "@/hooks/useHamburgerIcon";
-import frame from "../../styles/frame.module.css";
+import frame from "../../../styles/frame.module.css";
 import Image from "next/image";
 import icon from "../../../public/Images/トナカイアイコン6 (1).png";
+import React from "react";
 import { HamburgerIcon } from "./Hamburger";
 
 export const Header = (props: TitleType) => {
