+++ src/pages/Home/index.page.tsx
@@ -1,0 +1,8 @@
+import Home from "@/components/pages/Home/Home";
+import React from "react";
+
+function Page() {
+  return <Home />;
+}
+
+export default Page;
