+++ src/pages/MyProfileScreen/index.page.tsx
@@ -1,0 +1,8 @@
+import MyProfileScreen from "@/components/pages/MyProfileScreen/MyProfileScreen";
+import React from "react";
+
+function Page() {
+  return <MyProfileScreen />;
+}
+
+export default Page;
