--- src/pages/ProductListScreen.tsx
@@ -1,16 +1,0 @@
-import { Footer } from "./components/Footer"
-import { Header } from "./components/Header"
-import Product from "./components/Product"
-
-const ProductListScreen = () => {
-   
-    return (
-        <>
-            <Header name="成果物一覧"/>
-            <Product/>
-            <Footer />
-        </>
-    )
-}
-
-export default ProductListScreen;
