+++ src/pages/ProductListScreen/index.page.tsx
@@ -1,0 +1,8 @@
+import ProductListScreen from "@/components/pages/ProductListScreen/ProductListScreen";
+import React from "react";
+
+function Page() {
+  return <ProductListScreen />;
+}
+
+export default Page;
