+++ src/pages/QualificationsScreen/index.page.tsx
@@ -1,0 +1,8 @@
+import QualificationsScreen from "@/components/pages/QualificationsScreen/QualificationsScreen";
+import React from "react";
+
+function Page() {
+  return <QualificationsScreen />;
+}
+
+export default Page;
