--- src/pages/components/Footer.tsx
@@ -1,12 +1,0 @@
-import frame from "../../styles/frame.module.css"
-
-export const Footer=()=>{
-    return(
-      <footer className={frame.footer}>
-        <p >
-         &copy; 2025 profile site
-      </p>
-      </footer>
-    )
-
-}
