--- src/pages/components/Product.tsx
@@ -1,14 +1,0 @@
-import HomeButton from "./HomeButton";
-import ProductList from "./ProductList"
-
-const Product = () => {
-    return (
-        <>
-            <ProductList />
-            <HomeButton/>
-        </>
-    )
-
-}
-
-export default Product;
